Audio: use u16 as type for 'channel_count'
The driver descriptors are stored in ROM, where 32-bit reads are slower
than 16-bit reads. Since 'channel_count' is a small number, it's better
to use the smaller and faster type.

+++ include/libsimplegba/hw/audio.h
@@ -44,7 +44,7 @@ extern const struct _AudioDriver {
     void (*panning)(i32 channel, i32 panning);
 
     // internal use
-    u32 channel_count;
+    u16 channel_count;
     i32 (*available_channel)(void); // returns -1 if none available
 } *_audio_driver;
